+++ lib/libesp32/berry/src/be_introspectlib.c
@@ -66,10 +66,25 @@ static void m_findmember_protected(bvm *vm, void* data)
 static int m_findmember(bvm *vm)
 {
     int top = be_top(vm);
-    if (top >= 2 && (be_isinstance(vm, 1) || be_ismodule(vm, 1) || be_isclass(vm, 1)) && be_isstring(vm, 2)) {
-        int ret = be_execprotected(vm, &m_findmember_protected, (void*) be_tostring(vm, 2));
-        if (ret == BE_OK) {
-            be_return(vm);
+    bbool protected = btrue; /* run protected, i.e. don't raise an exception if not found */
+    if (top >= 3) {
+        protected = !be_tobool(vm, 3);
+    }
+    if (top >= 2) {
+        if (protected && (be_isinstance(vm, 1) || be_ismodule(vm, 1) || be_isclass(vm, 1)) && be_isstring(vm, 2)) {
+            int ret = be_execprotected(vm, &m_findmember_protected, (void*) be_tostring(vm, 2));
+            if (ret == BE_OK) {
+                be_return(vm);
+            }
+        } else {
+            /* run unprotected */
+            if (be_getmember(vm, 1, be_tostring(vm, 2))) {
+                be_return(vm);
+            } else {
+                /* not found, return module 'undefined' */
+                be_getmodule(vm, "undefined");
+                be_return(vm);
+            }
         }
     }
     be_return_nil(vm);
@@ -143,6 +158,20 @@ static int m_getmodule(bvm *vm)
     be_return_nil(vm);
 }
 
+/* set or chang the cached value for the named module, this allows monkey patching. **USE WITH CARE** */
+static int m_setmodule(bvm *vm)
+{
+    int top = be_top(vm);
+    if (top >= 2) {
+        bvalue *v = be_indexof(vm, 1);
+        if (var_isstr(v)) {
+            be_pushvalue(vm, 2);  /* ensure the second arg is at top of stack */
+            be_cache_module(vm, var_tostr(v));
+        }
+    }
+    be_return_nil(vm);
+}
+
 /* checks if the function (berry bytecode bproto only) is hinted as a method */
 static int m_ismethod(bvm *vm)
 {
@@ -167,6 +196,7 @@ be_native_module_attr_table(introspect) {
     be_native_module_function("set", m_setmember),
 
     be_native_module_function("module", m_getmodule),
+    be_native_module_function("setmodule", m_setmodule),
 
     be_native_module_function("toptr", m_toptr),
     be_native_module_function("fromptr", m_fromptr),
@@ -184,6 +214,7 @@ module introspect (scope: global, depend: BE_USE_INTROSPECT_MODULE) {
     set, func(m_setmember)
 
     module, func(m_getmodule)
+    setmodule, func(m_setmodule)
 
     toptr, func(m_toptr)
     fromptr, func(m_fromptr)